+++ python/src/ouster/client/__init__.py
@@ -9,6 +9,8 @@ from ._client import MultipurposeIOMode
 from ._client import Polarity
 from ._client import NMEABaudRate
 from ._client import ChanField
+from ._client import UDPProfileLidar
+from ._client import UDPProfileIMU
 from ._client import SensorConfig
 from ._client import get_config
 from ._client import set_config